[groonga] Add runtime dependency on onigmo
It is intended that included onigmo is used as static library,
but it is build as shared library and used, but not installed.
It is caused due to LIBTOOL_OVERRIDE, generated `libtool` is
replaced with pkg's libtool for shared library.

It is hard to fix it.
But this package have configure option to build onigmo as shared
library and install.
Additionally, it is notead as TODO to support using system one.
Moreover, bundled one is exactly same as onriginal one.
So it is safe to use the shared library from "onigmo" package
to fix runtime dependency.