Add p5-Text-CSV-Simple 1.00
Parsing CSV files is nasty. It seems so simple, but it usually isn't.
Thankfully Text::CSV_XS takes care of most of that nastiness for us.
Like many modules which have to deal with all manner of nastiness and
edge cases, however, it can be clumsy to work with in the simple case.
Thus this module.
We simply provide a little wrapper around Text::CSV_XS to streamline
the common case scenario.