Update to 1.3.0
* Use FFmpeg 3

Changelog:
2016-02-16, Tuesday :: MPlayer 1.3.0 released

MPlayer 1.3.0 "worksforme" is out.

Less than a month after 1.2.1, we're at it again.

MPlayer 1.3.0 is compatible with the FFmpeg 3.0.x releases and (at the time of writing) with FFmpeg git. The tarball already includes a copy of FFmpeg 3.0, so you don't need to fetch it separately.

This release brings you some new codecs and formats, a lot of fixes, and many cleanups. It also includes all the enhancements and speed-ups from FFmpeg; check their changelog if you are curious about the details.

In addition to these, there were a lot of updates to accommodate the API changes made in FFmpeg. Some were simple renames, but others were quite invasive. None of them should have a user-visible effect, except maybe for some corner-cases in the channel order for multichannel files.