Import libdbusmenu-glib-12.10.2 as devel/libdbusmenu-glib.
A small little library that was created by pulling out some comon code out of
indicator-applet. It passes a menu structure across DBus so that a program can
create a menu simply without worrying about how it is displayed on the other
side of the bus.

This package provides GLib2 API.