Release 3.13.0:
Postpone I/O associated with TEMP files for as long as possible, with the hope that the I/O can ultimately be avoided completely.
Merged the session extension into trunk.
Added the ".auth ON|OFF" command to the command-line shell.
Added the "--indent" option to the ".schema" and ".fullschema" commands of the command-line shell, to turn on pretty-printing.
Added the ".eqp full" option to the command-line shell, that does both EXPLAIN and EXPLAIN QUERY PLAN on each statement that is evaluated.
Improved unicode filename handling in the command-line shell on Windows.
Improved resistance against goofy query planner decisions caused by incomplete or incorrect modifications to the sqlite_stat1 table by the application.
Added the sqlite3_db_config(db,SQLITE_DBCONFIG_ENABLE_LOAD_EXTENSION) interface which allows the sqlite3_load_extension() C-API to be enabled while keeping the load_extension() SQL function disabled for security.
Change the temporary directory search algorithm on Unix to allow directories with write and execute permission, but without read permission, to serve as temporary directories. Apply this same standard to the "." fallback directory.


Bug Fixes:
----------
Fix a problem with the multi-row one-pass DELETE optimization that was causing it to compute incorrect answers with a self-referential subquery in the WHERE clause. Fix for ticket dc6ebeda9396087
Fix a possible segfault with DELETE when table is a rowid table with an INTEGER PRIMARY KEY and the WHERE clause contains a OR and the table has one or more indexes that are able to trigger the OR optimization, but none of the indexes reference any table columns other than the INTEGER PRIMARY KEY. Ticket 16c9801ceba49.
When checking for the WHERE-clause push-down optimization, verify that all terms of the compound inner SELECT are non-aggregate, not just the last term. Fix for ticket f7f8c97e97597.
Fix a locking race condition in Windows that can occur when two or more processes attempt to recover the same hot journal at the same time.